+++ source/dotnet/domain-tests/Fixtures/CalculationFixture.cs
@@ -16,6 +16,9 @@ using Azure.Identity;
 using Azure.Messaging.ServiceBus;
 using Azure.Messaging.ServiceBus.Administration;
 using Energinet.DataHub.Wholesale.DomainTests.Clients.v3;
+using Energinet.DataHub.Wholesale.DomainTests.Fixtures.Configuration;
+using Energinet.DataHub.Wholesale.DomainTests.Fixtures.Identity;
+using Energinet.DataHub.Wholesale.DomainTests.Fixtures.LazyFixture;
 using Moq;
 using Xunit.Abstractions;
